+++ src/chart/options/config.cpp
@@ -33,6 +33,12 @@ std::string Config::paramsJson()
 		for (const auto &param : channelParams)
 			arr << "channels." + std::string{channelName} + "."
 			           + std::string{param};
+
+		if (channelName == "x" || channelName == "y")
+			for (const auto &param :
+			    getAccessorNames<AxisChannelProperties>())
+				arr << "channels." + std::string{channelName} + "."
+				           + std::string{param};
 	}
 	return res;
 }
@@ -112,6 +118,14 @@ void Config::setChannelParam(const std::string &path,
 
 	if (auto &&accessor = getAccessor<Channel>(property).set)
 		accessor(channel, value);
+	else if (auto &&axisAccessor =
+	             getAccessor<AxisChannelProperties>(property).set;
+	         (channelId == AxisId::x || channelId == AxisId::y)
+	         && axisAccessor)
+		axisAccessor(
+		    options.getChannels().axisPropsAt(
+		        channelId == AxisId::x ? AxisId::x : AxisId::y),
+		    value);
 	else
 		throw std::logic_error(
 		    path + "/" + value
@@ -130,6 +144,11 @@ std::string Config::getChannelParam(const std::string &path) const
 
 	if (auto &&accessor = getAccessor<Channel>(property).get)
 		return accessor(channel);
+	if (auto &&axisAccessor =
+	        getAccessor<AxisChannelProperties>(property).get;
+	    (id == AxisId::x || id == AxisId::y) && axisAccessor)
+		return axisAccessor(options.get().getChannels().axisPropsAt(
+		    id == AxisId::x ? AxisId::x : AxisId::y));
 
 	throw std::logic_error(path + ": invalid channel parameter");
 }